feat(appconfig): Constrain environments to a single deployment at a time
[Issue #29345]
------------------------------------------------------------------------

[Reason for this change]
------------------------------------------------------------------------
The current L2 AppConfig constructs do not have any guardrails that
prevent simultaneous Deployments to a single Environment.
This is not allowed, and will result in Cfn deploy-time conflicts.

[Description of changes]
------------------------------------------------------------------------
This commit adds a pair of new public methods to IEnvironment that
enable the addition of a new Deployment for a given IConfiguration.

It then updates the creation of new Deployments in ConfigurationBase
to utilize these new methods instead of the current resource creation.

These new methods interact with an internal queue.
This queue creates a chain of Cfn dependencies between Deployments
in order to enforce that only a single Deployment can be in progress
for the Environment at any given time.

[Description of how you validated changes]
------------------------------------------------------------------------
Added new unit test coverage.