chore(s3): upgrade S3BucketNotifications Lambda to python3.11 (#29506)
### Issue # (if applicable)

Closes D122674750.

### Reason for this change

This Lambda does not work in new regions.

### Description of changes

Current S3BucketNotifications Custom Resource Lambda `python3.9` uses `botocore` version `1.29.90` which does not have S3 endpoint information for new regions.

Upgrading to `python3.11` which has `botocore` version `1.30.1`. See more in D122674750.

Note that we cannot upgrade to `python3.12` because it's not available in all regions yet.

### Description of how you validated changes

N/A.

### Checklist
- [X] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*