docs(opensearchservice): change default value's description for `multiAzWithStandbyEnabled` (#29634)
### Reason for this change

Default value of the `multiAzWithStandbyEnabled` parameter will be set to true if the feature flag `multiAzWithStandbyEnabled` is enabled now.

However, the current description for the default value does not yet reflect that.

### Description of changes

Change the description for the default value with the feature flag.

### Description of how you validated changes

Nothing. It is just to change the description.

### Checklist
- [x] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*