feat(sns): add TracingConfig prop (#29783)
### Issue # (if applicable)

Closes #29714

### Reason for this change

Currently, to set the TracingConfig, it is necessary to configure it via L1.
So, add TracingConfig props to L2.

### Description of changes
added TracingConfig props to topic.ts, sns.test.ts, integ.sns.ts, and README.md for AWS SNS.

### Description of how you validated changes
I confirmed with unit test and integ test that it works as expected.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*