chore(cli-testing): add a retry for test
One of our tests can remove customPermissionsBoundary creates a policy using createPolicy.
Change to IAM policies/roles use eventual consistency. So, while the changes will show up right
away if we were to call an API to describe that policy/role, the updates may not have actually propagated
to all regions yet. This is likely the cause of the intermittant test failures for this test.

This change adds the eventually block and uses it to retry initial creation of this stack in the case
that the policy changes have not made it to the relevant region just yet.